Refactored the code, so it fits the specific domain of the project. Renamed classes to fit the domain, added new classes, as needed, to fit the domain
Renamed the Room clas to Region, and changed some of the properties, so they properly fit the domain of the project. Added State class, which represents the state the player lands in, in a specific region. Added TurnCounter, World and Invetory singleton classes, to handle different aspects of the application. Each class serves an unique role