implemented the logic for reading the regions data from a json file (region-data.json)
Implemented the logic for reading the data for each region from a json file. The game utilises a file reader class, to read the file, the class currently contains only one method, which is called ReadJsonFile. Implemented a class, that is responsible for getting the data read from the JSON file, mapping through a DTO and then parsing to the class in the application. Code is tested and it works