Finished the system for dynamic init of regions
Finished the system for dynamic init of regions. Created a class RegionsService, that is in charge of handling all operations related to the Region object. Currently the RegionService has one public method - InitialiseRegions, where the reading from file, initialisations and assignment of exits takes place, and a private method, that converts list of Regions to a dictionary, for use of the InitialiseRegions method.