First version of the logic, for player has two turns letf
Set up the method in the TurnCounter, that initialises the state of last two turns - the method is called InitLastTwoTurns. Also set up the reverse method, which un-setts the last two turns. Changed the login in the CheckCrisisCondition method part of the CpiTracker. Additinally implemented the method that checks for crisis state changes in the Game class.